layerdiff: small improvements
The main change here is the complete removal of .ensureParent and that
we now sort the delta entries. While sorting entries is not necessary
(as far as I'm aware, ordering only matters in tar archives if you have
two of the same file being modified) it ensures that
layerdiff.GenerateLayer is more deterministics than before.

Signed-off-by: Aleksa Sarai <asarai@suse.com>