image: layer: initial UnpackManifest implementation
This is code works in most cases. Unfortunately, the current
implementation does not handle hardlinks properly (it instead uses
symlinks). The reason for this is that we cannot be sure that the
hardlink entries will be hit before the actual entry we want to unpack.
There are several potential solutions, but they will be implemented
later.

Signed-off-by: Aleksa Sarai <asarai@suse.com>