unpack: set root {a,m}time to unix epoch
The reason for this is that images like ubuntu will extract correctly
but they don't include an entry for '.'. A full solution for this would
include having a default set of metadata that we apply to '.'. But this
works for now.

In particular, it means that you can run gomtree on two different
bundles extracted at different times and be sure that they are truly
comparable.

Signed-off-by: Aleksa Sarai <asarai@suse.com>