vendor: update go-mtree and switch patching method
Rather than vendoring my fork of go-mtree, instead maintain a set of
patches that are applied to go-mtree (which is currently on master). The
hope is that once go-mtree releases the next version all of the issues
will no longer be necessary to be patched.

The only patch currently carried is:
 + gomtree-0001-entries-prepend-all-hierarchies-with-keyword-metadat.patch

Signed-off-by: Aleksa Sarai <asarai@suse.com>