pkg: implement unpriv.* as a library
This code came from my initial attempts in vbatts/go-mtree#96, but
refactored to be a library. The hope is that this will be vendored by
go-mtree so that we can share the same magical unpriv.* code.

But even if this doesn't land in go-mtree, we can just patch our version
to use it.

Signed-off-by: Aleksa Sarai <asarai@suse.com>