cmd: *: unify --image and --from
The usage for --image and --from was quite frustrating and also was
confusing (since the image path and the tag are quite related concepts).
So switch to an --image <path>[:<tag>] layout.

In addition, a bunch of changes to help pages, categorisation of
commands and --layout flags (which are added to differentiate between
"layout" commands and "image" commands) were done just to make
everything a bit more consistent.

Signed-off-by: Aleksa Sarai <asarai@suse.com>