image: generator: implement annotation generation to runtime
It's important the we set annotations in the resultant runtime config --
otherwise we could be dropping important configuration information. For
some reason this information is duplicated in two places in the
image-spec so we just pick an arbitrary ordering (that will need to be
clarified in a proposal to opencontainers/image-spec#479).

Signed-off-by: Aleksa Sarai <asarai@suse.com>