oci: cas: fix 'invalid cross-device link' error
Fix some odd errors that happened only if you were on btrfs, or
otherwise had /tmp and (the test directory) on different filesystem
mounts. In essence what happened is that e.temp would always be "" which
would cause ioutil.TempFile to create a temporary file under /tmp which
then would cause issues with os.Rename if you didn't also have your
image underneath /tmp.

The reason the tests used to work is because you'd get a different error
(cross-device link) because we were remounting the image directory (and
the fact it was on a different mount was the reason the error was
happening).

Fixes: 362fedad66da ("oci: cas: allow opening read-only CAS images")
Signed-off-by: Aleksa Sarai <asarai@suse.com>