oci: cas: fix *Readonly tests
The previous code was causing errors for other reasons (related to
cross-device linking that was caused by other issues). It turns out that
in order to create an "ro" bindmount you need to mount and then remount
it (not sure why, but that's what an strace of mount(1) told me).

Fixes: e6aac96be0a2 ("oci: cas: add readonly tests")
Signed-off-by: Aleksa Sarai <asarai@suse.com>