oci: layer: add xattrs to new layers
By necessity we have to ignore some xattrs (such as security.selinux),
but hopefully this is a rare case of things we need to ignore -- I don't
really want to have to maintain a set of xattrs that are and are not
safe to handle.

This also adds the "xattrs" keyword to the default set when doing
unpack. Otherwise we won't get an InodeDelta for the file.

Fixes: cyphar/umoci#52
Signed-off-by: Aleksa Sarai <asarai@suse.com>