cmd: hardcode go-mtree keywords
Because vbatts/go-mtree#87 wasn't merged upstream (and I'm not sure if I
want to carry that patch for other reasons), we have to just hardcode
what go-mtree keywords we're going to use.

Is it annoying that we can't use dh.UsedKeywords()? Yes. But there's not
much else we can do.

Signed-off-by: Aleksa Sarai <asarai@suse.com>