vendor: update vbatts/go-mtree@5685419c3e2d20ce9e4c556d7fa6ff2cb985b3b4
This also includes the removal of the following libraries:
 * golang.org/x/crypto

And the patches that used to be necessary are now applied upstream. In
addition, this update includes a fix to how unicode handling is done
with go-mtree. Hopefully this code will eventually be moved to a
separate library, and be completely rewritten (with a proper test
suite).

Ref: vbatts/go-mtree#118
Fixes: openSUSE/umoci#80
Signed-off-by: Aleksa Sarai <asarai@suse.de>