oci: cas: *ispec.Descriptor -> ispec.Descriptor
The previous API didn't make much sense, and was quite confusing when
actually trying to use it (nil would always cause errors, and the
intention of the call is not to modify the arguments).

In addition, clean up a lot of blob.Data casting and generation. It's a
bit uglier now, but in order to drop the pointers everywhere we need to
handle this case now.

This is part of the preparation of putting oci/cas into an OCI
repository.

Signed-off-by: Aleksa Sarai <asarai@suse.de>