.: remove LICENSE boilerplate
Remove LICENSE as it confuses GitHub's automated license detection. The
header in every file describing the license is sufficient. Not to
mention that there's also a COPYING in the root of the source tree with
the full license text.

I could rant for several hours about how insanely stupid GitHub's
license detection code is, and that it's just ridiculous that they don't
allow people to add boilerplate to projects -- because it makes it
impossible to have a standard way of telling contributors what the
project's full license terms are.

For instance, if you have a project under a GPLvX-or-later license, it
is not sufficient to just have a COPYING with the GPLvX license text --
you also need to include in the license header that it is licensed under
"GPLvX *or later*". Yes, you can (and should) include the header in all
source files, but that means that someone reviewing the license of your
project would have to *read every license header in the project* to be
sure of your project's licensing stance wrt. the or-later clause in the
GPL.

What insanity. Why didn't they just do what GNU Savannah did and allow
users to specify their license manually?

Signed-off-by: Aleksa Sarai <asarai@suse.de>