oci: split out gc and blob walking to casext
Garbage collection, blob auto-detection and walking are all built on top
of the CAS abstraction and don't belong in the same place as the
implementation or Engine interface.

In addition, Instead of making cas.Engine an argument to every extension
function, use the embedded struct method of adding extensions to an
interface.  This also provides a template for how other extensions
(third-party or otherwise) could be written.

Signed-off-by: Aleksa Sarai <asarai@suse.de>