test: rework setup_testdir to clean up properly
Previously we dropped bundle cleanups because rm(1) is not sophisticated
enough to know how to remove rootless rootfs-es. But this causes issues
with Travis (which has a cap on root filesystem size) so hack around the
rm(1) issues with a chmod 0777 -- which hopefully should handle all of
the important cases.

We have to track everything created with setup_testdir in a separate
manifest file because it turns out that global variables are hard to
modify in bash.

Signed-off-by: Aleksa Sarai <asarai@suse.de>