oci: casext: add dummy refname wrappers
This is just a temporary implementation which will be replaced with the
switch to v1.0.0-rc5. Note that this interface is *very* likely to
change in the future due to the arbitrary nesting requirements of
v1.0.0-rc5.

In addition, move the tests for references to oci/casext (using the new
interfaces). On paper this means that both the old and new interfaces
should still pass all of the oci/casext tests.

Signed-off-by: Aleksa Sarai <asarai@suse.de>