ci: fix EPERM coverage file
This recently broke on Travis, and it appears that there's some
mysterious permission issue happening with the interactions between
Docker and the coverage file.

Really we should make the CI not use Docker at all, but the package
requirements are forcing us to use CI for now. Hopefully we can improve
this in the future.

Signed-off-by: Aleksa Sarai <asarai@suse.de>