oci: cas: move PutBlobJSON to casext
Since PutBlobJSON is not actually an underlying property of the CAS, it
makes more sense in casext to remove the amount of things that CAS
implementations need to deal with. In addition it allows us to switch to
deterministic JSON encoding in the future without changing the CAS
implementations.

In addition, move the corresponding PutBlob tests into casext so that
the tests make more sense. The readonly and readwrite methods of testing
the API guarantees is not as nice as it should be at the moment, and
should be improved by breaking up the CAS "types" in the future (sort of
like io.Reader and io.Writer).

Signed-off-by: Aleksa Sarai <asarai@suse.de>