repack: add --no-mask-volumes and --mask-path
The reason for the split is that some users may wish to mask arbitrary
paths that have not been set with --config.volumes (which therefore
requires --mask-path). We make --no-mask-volumes an opt-out option
(despite the backwards compatibility problem) because it's a security
feature that really shouldn't be opt-in. I don't anticipate this will
break current users of KIWI (and future users will hit this and realise
it's a feature). However, this needs to be toggle-able because:

1. It is a backwards-incompatible change to the command-line API. Since
   KIWI has its own opinions about configuration and its relation to
   file unpacking, so older users of umoci need a way to switch to the
   old semantics.

2. Some users don't want to follow the suggestion of the spec that
   volumes should not be included in layers. While it's not clear when
   this would be a good idea (--no-mask-volumes is a security feature
   against credential exposure), we must retain it as an option. If
   nobody ends up using it we can remove the ability to disable it.

Signed-off-by: Aleksa Sarai <asarai@suse.de>