*: update to image-spec@v1.0.0-rc5
This requires some very significant changes to various portions of
umoci. At this point, umoci supports all of the necessary bits for
"standard" images (old-style structure with new layout). There are some
leftover bits in mutate/ as well as in the UX that are blocking
supporting the largest possible range of control over the structure of
OCI images generated and mutated by this library. These will be fixed in
future revisions before the next release.

Signed-off-by: Aleksa Sarai <asarai@suse.de>