casext: change Walk to use DescriptorPath
Without this information, it's impractical for users of casext.Walk to
reconstruct the descriptor path from the root. This is necessary for a
variety of uses, but the most prominent one is so that mutate/ is able
to ensure that it has replaced all necessary blobs in the descriptor
path so that future de-referencing of the blob will not break.

As an aside, casext.Walk really could use some proper tests.

Signed-off-by: Aleksa Sarai <asarai@suse.de>