mutate: handle DescriptorPath properly
Now that we have DescriptorPath supported everywhere else, switch over
mutate so that we are actually mutating a DescriptorPath. The key trick
is that we have to walk up the reference path in reverse and replace all
references to the child in the parent (which allows us to re-create the
walk from the root by only modifying the descriptor we care about).

This also adds unit tests to make sure that path rewriting actually
works as intended. In the future this will need to be verified through
integration testing with {unpack,repack}.

Signed-off-by: Aleksa Sarai <asarai@suse.de>