*: update to v1.0.0 of {runtime,image}-spec
On paper this change should not affect anything, but we needed to add a
patch in order to fix a late change in runtime-spec that broke the
runtime-tools/generate build. It also required updating runtime-tools so
that the verification didn't fail (image-tools is still too lax to
notice the issues).

Signed-off-by: Aleksa Sarai <asarai@suse.de>