layer: always set bundle to 0700 on creation
This protects against attacks where an unsafe image is extracted into a
location where an unprivileged user could potentially access unsafe
setuid binaries (or unsafe inodes). This protection is very easy to
disable by a user of umoci, but the intention is to protect against
lowest-common-denominator attacks.

This is a slightly breaking change, but as it is a security issue (which
was previously documented) we have made the change anyway.

Signed-off-by: Aleksa Sarai <asarai@suse.de>