extract: handle parent directory xattr restores
Previously, our "parent directory restore" handling (used to make sure
that we don't break the mtime of parent directories and other such
things) didn't handle preservation of xattrs properly. This was caused
by os.Lstat not reading the xattrs from the path (which then lead to the
xattrs not being present in the tar.Header that is used for
restoration). The fix for this is quite simple, just add the xattrs
manually.

Signed-off-by: Aleksa Sarai <asarai@suse.de>