*: add ability to use a shared CAS directory
umoci currently always expects blobs to live within the image layout
it's working on, in the aptly-named "blobs" subdirectory. However, the
OCI specification [1][1] permits this directory to be empty and for
implementations to look in other locations for referenced blobs.
This is particularly useful when working with multiple image layouts and
wanting to share blobs between them.

This patch adds a `--shared-cas` flag to all umoci commands that work
with image layouts. The flag expects a directory to be passed, and if
set, this directory will be used for all blob operations.

In the first implementation, we extend the `dirEngine` implementation of
the CAS interface to accept another (optional) `sharedCasPath` field.
Analogously to the CLI flag, if this field is supplied, it will be used
internally by the dirEngine for all blob-related operations.

[1]: https://github.com/opencontainers/image-spec/blob/7c889fafd04a893f5c5f50b7ab9963d5d64e5242/image-layout.md#blobs

Signed-off-by: Jonathan Boulle <jonathanboulle@gmail.com>