Makefile: disable SELinux for tests in Docker
With the current Makefile, the Docker variants of the tests don't work
out of the box on SELinux-enabled systems because the volume mount
(namely, the source directory) isn't labelled correctly. This patch
simply disables SELinux labelling for these particular containers,
as the easiest way to mitigate this issue. (This seems preferable to
using the `:z` or `:Z` option to the Docker volume mount, which would
forceably relabel the directory every time the Docker command is run.)

Signed-off-by: Jonathan Boulle <jonathanboulle@gmail.com>