Makefile: remove volume mount of source directory
With the current Makefile, the Docker variants of the tests don't work
out of the box on SELinux-enabled systems because the volume mount
(namely, the source directory) isn't labelled correctly.

However we're already copying in the source, and the volume mount (over
the top of it) was only really used for `make shell`. Since this is an
uncommon use case, let's just drop the volume mount completely.

Signed-off-by: Jonathan Boulle <jonathanboulle@gmail.com>