Makefile: add option to specify individual test(s)
Running all the integration tests can be a bit tedious when a developer
is iterating on a particular subcommand. This adds the ability to
specify a subset of tests by passing the `TESTS=` argument to the
`test-integration` and `local-test-integration` Make targets. This
argument expects the name of one or more tests, each of which is
expected to match a .bats file in the `tests/` directory. If unset, we
continue to default to running all tests.

Examples:
  $ make TESTS="gc" test-integration
  $ make TESTS="unpack repack" local-test-integration
  $ make test-integration

Signed-off-by: Jonathan Boulle <jonathanboulle@gmail.com>