oci: tar extract: remove non-dir targets on unpack
In order to work around cases where we are replacing a file we don't
have write access to in unprivileged mode (as well as to clarify the
hardlink semantics of unpacking), always remove the target path (unless
it's a directory) before unpacking over it. As we always re-apply
metadata this should not cause any issues.

Signed-off-by: Aleksa Sarai <asarai@suse.de>