oci: tar generate: disallow .wh. prefix files
The OCI specification makes it clear that we MUST NOT allow users to
have a rootfs that contains '.wh.'-prefixed files. While we cannot
prevent them from creating such a rootfs, we can abort when we see it.
Also add a test to ensure that this doesn't regress.

In addition we error out if we see opaque whiteouts. This is a temporary
patch in this series, and will be replaced with full opaque whiteout
support in a future commit.

Signed-off-by: Aleksa Sarai <asarai@suse.de>