test: add tests for user.rootlesscontainers
Add some tests for user.rootlesscontainers to make sure that owner
propagation is handled correctly when the "virtual" owner is modified in
rootless mode. Unfortunately we don't have any bash helpers for
rootlesscontainers.proto, so we have to compare the xattrs to
pre-computed hex values (this will likely break on upgrading the
protocol).

Signed-off-by: Aleksa Sarai <asarai@suse.de>