pkg: unpriv: add Walk implementation
This adds a filepath.Walk implementation to unpriv, that works
unprivileged. At the moment the WalkFunc is Wrap'd internally, though
this results in the function being called more than once if there's a
permission error. This might not be ideal for all users, but it is more
transparent I guess?

Also add Walk to the fseval interfaces so we can use Walk in oci/layer
for the purposes of handling opaque whiteouts.

Signed-off-by: Aleksa Sarai <asarai@suse.de>