oci: tar extract: enable some rootless tricks in user namespace
Some rootless tricks we do (such as faking block devices) are useful in
user namespaces, while others (ignoring Lchown) are not. In the future
the way we handle this will probably change (especially with the advent
of seccomp-userspace traps which would allow for mknod(2) emulation in
userspace -- removing the privilege problem), but currently we enable
some tricks if we are detected as being in a user namespace.

It should be noted that user namespace checks are very fickle. There
isn't any way to guarantee you are not in a user namespace without
trying something like mknod(2) and noticing it fails when it shouldn't
(and even that could be blocked by other things). So the check we do
(see if /proc/self/uid_map has the same value as the root user
namespace) is 99% accurate though it can be tricked quite trivially.

We don't have tests for this yet, because we don't really have the
ability to test that within our test setup.

Reported-by: Serge Hallyn <serge.hallyn@ubuntu.com>
Signed-off-by: Aleksa Sarai <asarai@suse.de>