oci: casext: extend scope of valid reference names
Previously we only supported a very limited subset of reference names
(as per the specification), but the restrictions have been expanded to
allow for any reference of the form

    refname   ::= component ("/" component)*
    component ::= alphanum (separator alphanum)*
    alphanum  ::= [A-Za-z0-9]+
    separator ::= [-._:@+] | "--"

Add support for this newly expanded set of reference names. In addition,
refactor the checks for whether a reference name is valid.

Signed-off-by: Aleksa Sarai <asarai@suse.de>