hack: remove opencontainers/runtime-tools patch
This one is quite ugly because it requires pulling in a bunch of random
projects because of how badly organised the validate.CapValid
dependency has been structured. I have proposed fixing this upstream,
but there doesn't appear to be much interest in actually fixing this
issue.

I've also included some updates with this patch -- and with this fix we
no longer patch any upstream projects.

Signed-off-by: Aleksa Sarai <asarai@suse.de>