oci: ignore system.nfs4_acl and extend forbidden-xattr handling
It turns out that system.nfs4_acl is yet another xattr that we shouldn't
be touching, but also that we should not be clearing (while we have
permission to do so, clearing it results in NFS permissions breaking).
So as an extension we now now use ignoreXattrs for both unpack and
repack operations.

Signed-off-by: Aleksa Sarai <asarai@suse.de>