vendor: switch to github.com/rootless-containers/proto@v0.1.0
This obsoletes the need for a dedicated pkg/rootlesscontainers-proto
(because I effectively just copied the sources). This allows us to
increase our test coverage as well as remove the need to keep
rootlesscontainers.proto in sync with upstream (we get that for free
because we now use the canonical repository for the .proto).

Signed-off-by: Aleksa Sarai <asarai@suse.de>