cmd: add 'umoci raw add-layer'
This is a very simple helper function that allows users to "BYO" layers
that were generated elsewhere. It should be noted that this does
slightly hardcode tar archives into umoci's interface more than I'd
like, but I'm sure we can work around that if we have to in the future.

Signed-off-by: Aleksa Sarai <asarai@suse.de>