test: switch to more safe BUNDLE boilerplate
The previous usage of $BUNDLE_{A..Z} would result in copy-paste errors
quite often (making the tests not actually test what they should). This
is quite easily solved by always using $BUNDLE (except in tests that
need to reference previous bundles).

In addition, this fixes the --keep-dirlinks tests to test all of the
important features (like nesting resolution as well as loop detection
and breakout).

Signed-off-by: Aleksa Sarai <asarai@suse.de>