unpack: fix symlink loop handling in --keep-dirlinks
Previously the symlink loop handling didn't really handle symlink loops
correctly (it also didn't really handle nested symlinks properly either
-- the path was wrong) so correct this mistake.

In addition some of the other weirdness around --keep-dirlinks was fixed
(like not removing the existing object if we decide it isn't a dirlink
-- in those cases you _need_ to remove it otherwise extraction will fail
with ELOOP).

Tests for this are added in a subsequent patchset (along with fixes to
the existing test).

Signed-off-by: Aleksa Sarai <asarai@suse.de>