umoci repack: create simplified whiteouts
This change ultimately makes our images a little bit smaller by no
longer whiting-out every sub-entry when a directory has been deleted.
While this is a net benefit, the cause of this change is a Docker
breakage (which arguably added a restriction that is contrary to the OCI
specification and thus it is definitely *their* bug). But to play nice
(and to take the opportunity to improve some images) we can fix it in
umoci.

Signed-off-by: Aleksa Sarai <asarai@suse.de>