umoci insert: extend to support whiteouts
Previously umoci-insert(1) was quite limited in that it would only
insert a given file or directory without having any configurability.
Something users might want is the ability to delete a path without an
unpack-repack cycle. Or they want to replace the contents of a directory
rather than merge them.

To this effect, add --whiteout and --opaque which provide the
corresponding features. --whiteout just adds a whiteout at the target
path, while --opaque adds an opaque whiteout for the target before doing
the copy.

You _can_ use --whiteout and --opaque together but it's basically
useless, and Docker might complain for no good reason.

Signed-off-by: Aleksa Sarai <asarai@suse.de>