oci: layer: skip forbidden xattrs if unchanged
It turns out that since SELinux auto-sets labels for every directory
created, 6fd1e0e62645 ("oci: ignore system.nfs4_acl and extend
forbidden-xattr handling") was far from sufficient to fix the overall
issue. In particular, restoring of parent metadata broke on such systems
because we try to re-apply the metadata we saw *on-disk*.

Erroring out because a forbidden xattr was on-disk is silly (and is
counter to what we actually want to do), so instead we can only error
out if we've been asked to change the xattr from the on-disk value. This
avoids us having to special-case restoreMetadat with on-disk data -- and
instead we can make it slightly more general. This does mean that images
which container "security.selinux" set "just right" will now be ignored,
I think that this is a small price to pay.

Signed-off-by: Aleksa Sarai <asarai@suse.de>