umoci: make --rootless hint louder
We added a hint if a user got EPERM to use --rootless to try to fix it
in a9ebb42831ad ("cmd: give 'helpful' hint about --rootless"). However,
since then we changed the logging to be a bit less noisy and this patch
was lost in the changes -- so we need to make this a bit louder to make
sure it's visible by default.

Reported-by: Vincent Batts <vbatts@hashbangbash.com>
Signed-off-by: Aleksa Sarai <asarai@suse.de>