mutate: fix Go 1.11 builds
In Go 1.11, the semantics of archive/tar were changed in a very worrying
way[1]. hdr.Typeflag values of TypeRegA are now silently converted to
TypeReg, which results in changes in the hashes of said objects. Luckily
this didn't affect the core oci/ code because we explicitly used TypeReg
-- but it did affect one of our tests.

This is another justification for pushing through with OCIv2, because
depending on a format that has no guaranteed canonical format is quite
dangerous. I do find it somewhat silly that a change like this was made
for such a minor reason (to better conform with one sentence of the PAX
specification -- in a format where nobody follows the rules). But I do
understand why they'd want to do that.

[1]: https://go-review.googlesource.com/c/go/+/85656

Signed-off-by: Aleksa Sarai <asarai@suse.de>