oci: handle Go 1.10 change to FileInfoHeader
Go 1.10's archive/tar decided to add Uname and Gname to archives by
default[1], which obviously is a bit of an issue for us if we want to be
reproducible across Go versoins. To counter-act this we explicitly set
them all to the zero value.

To avoid chasing shadows in the future, we'll add some regression tests
to make sure that Go's archive/tar doesn't regress us again.

[1]: golang/go@0564e304a6ea ("archive/tar: populate uname/gname/devmajor/devminor in FileInfoHeader")

Signed-off-by: Aleksa Sarai <asarai@suse.de>