test: add archive/tar regression test
Normally regression tests are for your own project, but (as seen
throughout umoci's history) we are currently acting as one of the few
regression tests of many other projects. In the past we've found plenty
of Go bugs and bugs in many other OCI tools, but this was usually due to
some pretty blatant problems.

Recently though there have been a slew of increasingly subtle Go
changes. Unfortunately it's not possible for us to be certain of whether
a Go compiler update will break our reproducibility between versions,
but we should at least be able to smoke test it somewhat.

Here's hoping I never have to see this test again.

Signed-off-by: Aleksa Sarai <asarai@suse.de>