*: add CODE_OF_CONDUCT and GOVERNANCE
The openSUSE Code of Conduct is fairly reasonable, and was already
effectively "the law of the land" since we are an openSUSE project. This
just codifies it, and makes it clear that the Board is the "appellate
court" for any violations.

In addition, add a basic BDFL governance model (but with an "escape
hatch" to allow for appeals to the Board in *extreme* cases).

Signed-off-by: Aleksa Sarai <asarai@suse.de>