hack: release: re-license under MPLv2
Since this script is not actually used in the final binary deployment
(it's just a useful script), limited file-based copyleft isn't an issue
-- and this license has the benefit that it's compatible with both
Apache 2.0 and GPLv2.

As can be seen in the git history, I am the sole author of this file and
thus there is no issue with relicensing it.

Signed-off-by: Aleksa Sarai <asarai@suse.de>