layer: specify "type: bind" in config.json for bind-mounts
Due to an oversight in runc, setting a type other than "" or "bind" will
result in weird errors (especially when dealing with file-based
bind-mounts).

To get around this, we can simply set an empty type. This oversight is
being dealt with upstream, but won't help old runc releases (so we need
this workaround -- and we use "bind" to work with much older releases as
well).

Signed-off-by: Aleksa Sarai <asarai@suse.de>