gc: drive gc using digests instead of image tags/names
garbage collection enforces an implicit policy that only images with
names and tags must be kept and the rest is fair game to prune.
However, as per the OCI spec [1], "annotations" field is optional.

This patch uses only digests to determine reachability.
If a particular image needs to be garbage collected then remove the
image from the index.json and run gc.

Note that this is a behavior change in that some images which were
pruned earlier will now not be.

Signed-off-by: Ramkumar Chinchani <rchincha@cisco.com>

References:
[1] https://github.com/opencontainers/image-spec/blob/master/manifest.md#image-manifest-property-descriptions