site: make site mirror in-repo documents
This removes the need for two README-like markdown files to maintain,
and allows us to remove the silly hacks from hack/publish-site.sh. We
need to define aliases for all of the key site pages, but that isn't too
much of a pain to do. We also move the keyring file to the root of the
repo, to make it easier to find.

Note that we cannot add front-matter to the README.md because of a bug
in Hugo[1].

[1]: https://github.com/gohugoio/hugo/issues/7398

Signed-off-by: Aleksa Sarai <asarai@suse.de>